+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
-* 0.1.0 (2023-11-16) Initial release
-* 0.1.1 (2023-11-16) Remove null configuration file; fix bad json syntax
+* 0.1.0 (2023-11-16): Initial release
+* 0.1.1 (2023-11-16): Remove null configuration file; fix bad json syntax
   highlighting on readme
+* 0.2.0 (2023-11-17): Add doctest; add `emphasis_strong` double style; add
+  more strong/emphasis examples to `input.md`; fix issue with strong code double
+  style; improve the double style algorithm; fix changelog; improve readme